Definitions inherit fields from STI parent class too
When `scoped_search` is called on a parent and then its subclass, a new
`Definition` will be created and stored on the subclass that inherits
and merges fields with its parent.

The named scope will only be registered on the parent, but the most
specific definition from the subclass is used when the scope is called.
This allows `Subclass.search_for` to support both parent and subclass
fields, but `Parent.search_for` to only support parent fields.

Previously fields registered on the subclass may have used the parent's
definition causing errors finding relations as the wrong class was used.

Only works with STI parents/subclasses, not regular inheritance or
abstract classes as table and column information can safely be retrieved
against either the parent or subclass with STI.

Fixes #135