Support inheritance of fields with abstract classes
Field has been split into two parts: FieldDefinition to store static
configuration supplied by the user, and Field to contain behaviour that
relies on the class being scoped (e.g. #to_sql, relations, or column
types).

FieldDefinition#to_field(a_klass) returns a Field instance, so a
definition can be converted to the Field. This means that:

    FieldDefinition + (klass < ActiveRecord::Base) = Field

The class is passed now into QueryBuilder from the named search scope
and used to generate the Field instances where needed.

This enables support for definitions to be set on abstract or parent
classes and subclasses to be queried, as the table name will vary
between parents and subclasses. The STI case is simpler as using the
`Parent.table_name` or column information works for all subclasses.